requirements.txt
web3==5.30.0
python-telegram-bot==13.13